requirements.txt
annotated-types==0.7.0
anyio==3.7.1
click==8.2.0
colorama==0.4.6
docopt==0.6.2
fastapi==0.109.1
h11==0.16.0
httptools==0.6.4
humanize==4.12.3
idna==3.10
netaddr==1.3.0
openvpn-api==0.3.0
openvpn-status==0.2.2
psutil==7.0.0
pydantic==2.4.2
pydantic_core==2.10.1
PyQt6==6.9.0
PyQt6-Qt6==6.9.0
PyQt6_sip==13.10.0
python-dotenv==1.1.0
PyYAML==6.0.2
six==1.17.0
sniffio==1.3.1
starlette==0.35.0
typing_extensions==4.13.2
uvicorn==0.23.2
watchfiles==1.0.5
websocket-client==1.8.0
websockets==15.0.1
